+++ Test/GuidEx.Tests.ps1
@@ -1,10 +1,10 @@
 BeforeAll {
     # Import the module, then use InModuleScope for private functions
     # see: https://pester.dev/docs/usage/modules#testing-private-functions
-    Import-Module GUIDEx -Force
+    Import-Module "$PSScriptRoot\..\GUIDEx.psd1" -Force
 }
 
-Describe "Deprecated Functions" {
+<# Describe "Deprecated Functions" {
     Context "Convert-UUID" {
         It "Should return a GUID" {
             $guid = Convert-UUID ([guid]'906F89D8-00DA-3583-8057-A7425FB9FDD4') -Reverse
@@ -35,7 +35,7 @@ Describe "Deprecated Functions" {
             $flippedBytes | Should -Be $expectedBytes
         }
     }
-}
+} #>
 
 Describe "Public Support Functions" {
     Context "Convert-UUIDSquished" {
@@ -44,18 +44,26 @@ Describe "Public Support Functions" {
             $guid | Should -BeOfType [guid]
         }
 
-        It "Should convert squished UUID to standard GUID format" {
-            $squished = "01020304-0506-0708-090a-0b0c0d0e0f10"
-            $guid = Convert-UUIDSquished $squished
-            $guid.ToString() | Should -Be "01020304-0506-0708-090a-0b0c0d0e0f10"
+        It "Should return a string with -Format" {
+            $guid = Convert-UUIDSquished "0102030405060708090a0b0c0d0e0f10" -Format n
+            $guid | Should -BeOfType [string]
+            $guid | Should -Be '403020106050807090a0b0c0d0e0f001'
         }
 
-        It "Should be idempotent with GUID input" {
-            $inputGUID = [guid]"01020304-0506-0708-090a-0b0c0d0e0f10"
-            $result1 = Convert-UUIDSquished $inputGUID
-            $result2 = Convert-UUIDSquished $result1
-            $result1 | Should -Be $result2
-            $result1 | Should -Be $inputGUID
+        It "Should convert squished UUID back to standard GUID format" {
+            $squished = "403020106050807090a0b0c0d0e0f001"
+            $expected = [guid]"{01020304-0506-0708-090a-0b0c0d0e0f10}"
+            $guid = Convert-UUIDSquished $squished
+            $guid | Should -Be $expected
+        }
+
+        It "Should be idempotent with pipeline support" {
+            $inputGUID = [guid]"{01020304-0506-0708-090a-0b0c0d0e0f10}"
+            $intermediate = [guid]"{40302010-6050-8070-90a0-b0c0d0e0f001}"
+            $result1 = $inputGUID | Convert-UUIDSquished
+            $result2 = $result1 | Convert-UUIDSquished
+            $result1 | Should -Be $intermediate
+            $result2 | Should -Be $inputGUID
         }
 
         It "Should be idempotent with string GUID input" {
@@ -78,20 +86,33 @@ Describe "Public Support Functions" {
         }
     }
     Context "Get-UUIDFromNamespace Function" {
-        It "Should return the expected v5 UUID for a domain" {
+        It "Should return the expected v3 UUID for a domain (UTF8)" {
+            # ref: https://uuid.ca/uuid3/
             # Arrange
-            $expectedGUID = [guid]'cfbff0d1-9375-5685-968c-48ce8b15ae17'
             $uriGUID = [guid]'6ba7b810-9dad-11d1-80b4-00c04fd430c8'
+            $expectedGUID = [guid]'907e1018-10ac-35d1-b62e-561c32033af0'
 
             # Act
-            $result = Get-UUIDFromNamespace -Namespace $uriGUID -Name 'example.com'
+            $result = Get-UUIDFromNamespace -Namespace $uriGUID -Name 'example.com' -Version 3 -Encoding UTF8
+
+            # Assert
+            $result | Should -Be $expectedGUID
+        }
+
+        It "Should return the expected v5 UUID for a domain (UTF8)" {
+            # Arrange
+            $uriGUID = [guid]'6ba7b810-9dad-11d1-80b4-00c04fd430c8'
+            $expectedGUID = [guid]'cfbff0d1-9375-5685-968c-48ce8b15ae17'
+
+            # Act
+            $result = Get-UUIDFromNamespace -Namespace $uriGUID -Name 'example.com' -Version 5 -Encoding UTF8
 
             # Assert
             $result | Should -Be $expectedGUID
         }
 
         It "Should return the expected v5 UUID for a Terminal profile" {
-            # see: https://learn.microsoft.com/en-us/windows/terminal/json-fragment-extensions#profile-guids
+            # see: https://learn.microsoft.com/en-us/windows/terminal/json-fragment-extensions#calculating-a-guid-for-a-built-in-profile
             # Arrange
             $internalNS = [guid]'{2bde4a90-d05f-401c-9492-e40884ead1d8}'
             $profileName = 'Ubuntu'
@@ -105,6 +126,7 @@ Describe "Public Support Functions" {
         }
 
         It "Should correctly convert v5 UUIDs for a Terminal fragment" {
+            # see: https://learn.microsoft.com/en-us/windows/terminal/json-fragment-extensions#generating-a-new-profile-guid
             # Arrange
             $terminalNS = [guid]'{f65ddb7e-706b-4499-8a50-40313caf510a}'
             $appName = 'Git'