+++ CHANGELOG.md
@@ -5,7 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [2.7.2] - 2025-10-27
+
+### Added
+- `New-WindowsTerminalUUID` function to generate official and fragment profile GUIDs.
+- Aliases `New-WindowsTerminalGUID` and `terminalguid`
+
+### Changed
+- `Convert-UUIDSquished` used on its own returns a GUID, while using `-Format` will output a string.
 
 ## [2.6.2] - 2025-09-15